Avoid address overlap when updating address configuration
It's not possible to replay bootstrap when the new admin network uses
the same network as the previous management network. The fix prevents
address overlap by removing the management address before adding new
ones.

Test plan
=========

[PASS] Execute bootstrap without admin network defined. Replay bootstrap
with the admin network set to the same network as the previous
management network, and update the management network to a different
value. Values must be updated in sysinv database and kernel.

[PASS] Initiate the subcloud bootstrap without defining the admin
network configuration. Interrupt the process after updating the sysinv
database, then configure the admin network by assigning it the same
network as the previous management network, while updating the
management network to a different value. Replay bootstrap using the
"dcmanager subcloud deployt bootstrap" command. The values must be
updated in both the sysinv database and the kernel.

[PASS] Execute bootstrap with the admin network configured. Replay
bootstrap after removing the admin network. Ensure that the values are
updated in sysinv database and kernel.

Closes-Bug: 2081057

Change-Id: I509860f4643ae787fbb996903080f4183c2ee167
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>