Fix for admin-keystone delay during legacy restore
This patch addresses a delay in the creation of the admin-keystone
service's PID during legacy restoration. The delay was due to the
timing of the admin-keystone service startup. To resolve this, a
direct query to Keystone was introduced, which speeds up the creation
of the admin-keystone PID. This ensures that the admin-keystone
service is available more quickly during the restoration process,
improving the overall efficiency and reliability of critical service
availability.

Test Plan:
[PASS] no issue with admin-keystone playbook task
[PASS] Legacy B & R AIO-DX
[PASS] Optimized restore AIO-SX
[PASS] Verify that dnsmasq config has correct interface after unlock

Closes-Bug: 2081158

Change-Id: Ic3ec45b9b80ca5e3bad41f3bae6a4e6327b1f3dd
Signed-off-by: rummadis <ramu.ummadishetty@windriver.com>