Remove memory requests from networking pods
Networking platform pods should not request memory resources,
otherwise it interferes with the ability of application pods
to reserve the full amount of memory that K8s says is available.

It is desired to explicitly set the memory requests to zero
in order to allow the application pods to reserve the
full amount of available memory.

The networking pods include:
 - multus-cni
 - sriov-cni
 - sriov-device-plugin

Testing:
- Build successful
- kubectl describe node to verify resource usage is 0
- Run on sriov enabled lab for sriovdp pod

Closes-Bug: 2080666

Change-Id: I57afc4585a34b6fb192faaebcbecc7fbea407356
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>