Avoid Keystone password rollback during enroll
The update of the OAM network as a pre-step before the subcloud
enrollment playbook triggers Puppet to apply class
"openstack::keystone::endpoint::runtime::post" based on the Hieradata
before the enrollment playbook, which may change the updated Keystone
users' passwords back, which causes Keystone authentication failure.

This commit checks the OAM update during the enroll-init process by
cloud-init finished and the manifest was applied before updating the
Keystone passwords to avoid this issue.

Considering the seed.iso always triggers a reboot before the enrollment
playbook, there's no need to recover the sysinv service as it is
always configured by the controller's manifest. To reduce the time to
wait for the puppet manifest applied of OAM update(around 5 minutes),
this commit also moves the sysinv password update back to the role
"update-keystone-data" to reduce the waiting time in the playbook.

This commit also adds a step to lock the controller to finish the
enrollment playbook, to ensure the controller can be reconfigured from
the same state as post-bootstrap.

Test-plan:
1. Passed - Test subcloud enrollment on an AIOSX controller-0,
verify the playbook succeeded and the controller-0 changed to locked
status.
2. Passed - Test subcloud enrollment on an AIODX controller-0,
verify the playbook succeeded and the controller-0 changed to locked
status.
3. Passed - Introduce a failure after updating the sysinv password,
verify the retry of the enrollment succeeded.

Closes-bug: 2078458
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: Icd96dc633d4293283597edf451cf5ef8d8f00319