Fix blkid command in wipe_osds.sh
When the disk is blank (no gpt partition), wipping the osds
fails. This happens because the blkid command, used to check
whether the OSD is of type 'bluestore', returns exit code 2
when the disk label is unknown.

To fix this, the condition "|| true" was added to the end of
the command to ensure that the script continues to run.

Test Plan:
- PASS: BnR on AIO-SX with blank disk (unknown label)

Closes-Bug: 2080901

Change-Id: I61c12f59d1866999586adc085cf02183479eeb38
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>