Update Keystone passwords in a script in rehoming
The subcloud rehoming playbook costs a lot time for tasks that need
remote SSH access in a high-delay network environment. To reduce the
time for subcloud rehoming, this commit switches to using a script
to update the Keystone passwords. With this change, the total amount
of time spent on rehoming in the environment with 1200ms latency
on the OAM network is reduced from around 1hr55mins to 1hr30mins.

Further work remains to combine some ansible tasks into scripts to
reduce the times on remote SSH.

Test plan:
1. Passed - rehome a subcloud in a low-delay network environment.
2. Passed - rehome a subcloud in a long-delay network environment.
3. Passed - enroll a subcloud.

Closes-bug: 2080983

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I4a68c04a0e628cbd02a10101983e5dc11804f2fd