Set address facts for cluster-host and mgmt nets in separate tasks
It was observed that if cluster-host is dual-stack and mgmt is
single-stack the task "Set secondary cluster-host addresses for
subsequent tasks/roles" was also testing mgmt network's variable
controller_floating_address_secondary in the same task as the
variables for cluster-host, breaking execution.

This change separates the tasks.

Test Plan
=========

[PASS] Install AIO-DX with mgmt as dual-stack and cluster-host as
        dual-stack
[PASS] Install AIO-DX with mgmt as single-stack and cluster-host as
        dual-stack
[PASS] Install AIO-DX with mgmt as dual-stack and cluster-host as
        single-stack
[PASS] Install AIO-DX with mgmt as single-stack and cluster-host as
        single-stack

Story: 2011027
Task: 51046

Change-Id: I2c63252edbfe432594804d525016779c0f0466ed
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>