Certificate installation failure in bootstrap replay
This commit fixes the certification insallation failure that
is generated whe running bootstrap replay. When bootstrap is
replayed, the region name is updated in sysinv,conf but not in
openrc. The fix consists in keeping the region name defined in
openrc when bootstrap is replayed.

Test Plan:

PASS: Deploy an AIO-SX, apply the change to the playbook and
run the bootstrap playbook. After bootstrap is completed, run
the bootstrap playbook again. Verify that the playbook execution
finishes without any errors.

PASS: Deploy a system controller, apply the changes to the
playbook and run the bootstrap playbook. After bootstrap is
completed, run the bootstrap playbook again. Verify that the
playbook execution finishes without any errors.

PASS: Deploy a system controller, apply the change to the playbook
and run all steps to finish the configuration. After the host is
available, deploy a subcloud. Verify that the subcloud is correctly
deployed without any errors.

Closes-bug: 2081101

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: If2168ed7e0388bb9faf5315090fbb05bfe31f9d9