Handling unavailable release metadata
As the result of commit [1], the N-1 release metadata is placed
in /opt/software/metadata/unavailable directory. This commit
removes the unavailable directory if the subcloud is prestaged
for-install with N release (24.09) data because the N-1 release
metadata is irrelevant.

[1] https://review.opendev.org/c/starlingx/update/+/928282

Test-Plan:

Setup:
Bring up a system controller with 24.09.
Upload N-1 load to in the system controller.
Deploy one N release (24.09) subcloud and
one N-1 release (22.12) subcloud.
Prestage both subclouds with N release data.

PASS: Verify that the unavailable metadata directory
      exists on the N-1 subcloud.

PASS: Verify that the unavailable metadata directory
      does not exist on N subcloud.

Closes-Bug: 2081826

Change-Id: Ibbf8e75eaca696c7f24dab40827966e07718d096
Signed-off-by: sbhardwa <sharang.bhardwaj@windriver.com>