Enhance playbook to configure IPsec during upgrades
This commit enhances the solution to configure and enable IPsec per the
addition of tasks to verify if SAs are established between nodes and
a task to verify if strongswan service remains active on each node.

Test Plan:
PASS: Deploy a standard system and upgrade from stx 9 to stx 10.
      Observe that IPsec is configured and enabled on all hosts, SAs
      are established and nodes are reachable.
PASS: On a standard sytem, execute disable-ipsec.yml playbook and
      observe that IPsec is disabled on all nodes.
PASS: On a standard system, execute enable-ipsec.yml playbook and
      observe that IPsec is enabled and configured on all nodes.

Story: 2010940
Task: 51110

Change-Id: I2f4826e6d1ffcc8b45a7e332fec5b4448d7df004
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>