Fix dnsmasq FQDN resolution during restore
This change addresses an issue during the  restore
process where the internal FQDN (controller.internal) was
not being resolved, causing the all services to fail.
The fix involves killing the existing dnsmasq process,
restarting networking, and starting dnsmasq to ensure
proper resolution of internal FQDNs as the /etc and services
configurations are restored.

Test Plan:
[PASS] Verify controller.internal is pingable
[PASS] Restore playbook completed sucessfully

Closes-Bug: 2083059

Change-Id: I0f2e4a2de59b2f2f266e92bf32ef257a80dd4672
Signed-off-by: rummadis <ramu.ummadishetty@windriver.com>