Remove restart the software controller and agent
In the current prestage "for-sw-deploy" the process includes
restarting the software controller and agent, which is not
required.
This commit eliminates the unnecessary restarts of both the
software controller and agent.

Test-Plan:

Setup:
Bring up a system controller with 24.09.
Deploy one N release (24.09) subcloud
Upload and deployed a patch 24.09.1
Prestage subcloud for-sw-deploy with N release.

PASS: Verify prestage is successful.

Closes-Bug: 2083304

Change-Id: I394b5ce81cf21a962357882e8d704ecbaaaf2cab
Signed-off-by: sbhardwa <sharang.bhardwaj@windriver.com>