Fix restore on DX failing due DRBD puppet
The restore action fails due lack of formatation on
partition drbd9.

This error was caused by the variable restore_extra_items.
When the variable is empty a bunch of unwanted extra backup
data will be unpacked, including the flag node_ceph_configured,
preventing the formatting of the drbd9 (ceph-mon partition)

Test plan:
  - PASS: Test B&R on a AIO-DX successfully
  - PASS: Test a restore with the "-e '{"restore_extra_items":
        ["etc/platform/.node_ceph_configured"]}'" parameter
	and check if the error persists, ensuring the
	functionality of this parameter.

Closes-Bug: 2083606

Change-Id: I9a9582635a1a04c4584dc4aebe94272ff45275c1
Signed-off-by: Caio Correa <caio.correa@windriver.com>