Improve ostree configuration script for prestage
This commit improves the script that is responsible for
configuring the ostree repository on the subcloud.

When the prestage is called for-sw-deploy and the ostree
repo is required to be synchronized, the synchronization
script configures the repository on the subcloud to point
to the remote URL, which corresponds to the IP of the
system controller.

ostree uses the http protocol to synchronize the repository
via the command line. The ostree configuration URL is
mapped to the directory where ostree is located on the
system controller (/var/www/pages/iso/24.09/ostree_repo/).

To sync the repo, a valid system controller URL is required,
which can be over http or https.

It may happen that some subclouds do not have https enabled,
which caused an error when configuring the remote repository,
since the system controller only supports requests over https.

This change improves that mechanism to determine the
appropriate URL to configure the remote ostree reference
on the subcloud.

Test Plan:
Setup #1:
     - Bring up a 24.09 system controller with https enabled.
     - Bring up a 22.12p6 subcloud with https disabled.

PASS - Prestage the subcloud for software deploy.
     - Verify that the release is prestaged on the subcloud.

PASS - Enable https on the subcloud.
     - Prestage the subcloud for software deploy.
     - Verify that the release is prestaged on the subcloud.

Setup #2:
     - Upload and deploy patch 24.09.1 on SC.
     - Bring up a 24.09 subcloud with https enabled.

PASS - Prestage the subcloud for software deploy.
     - Verify that the 24.09.1 patch is prestaged on the
       subcloud.

PASS - Disable https on the subcloud.
     - Prestage the subcloud for software deploy.
     - Verify that the 24.09.1 patch is prestaged on the
       subcloud.

Story: 2010676
Task: 51132

Change-Id: Ib9700f48b93fca727f8b526878ffeec9b2b64527
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>