Install ssl cert during upgrade
In the upgrade to stx 10, systems will have https enabled during the
upgrade by the 'create-required-platform-certs.py' activation script.
In some scenarios where the active controller is too busy, the ssl
certificate generated to enable https takes too long to be installed
in the '/etc/ssl/private/' directory.

The lack of the expected certificate causes haproxy to fail once the
https reconfiguration is done, which by consequence triggers an
uncontrolled swact in DX systems. Controller-1 has the certificate
already, so the system is able to recover after swacting.

This review installs the ssl certificate as soon as it is issued by
cert-manager during the execution of 'create-required-platform-
certs.py', making the certificate available before the
reconfiguration to https.

Test plan:
PASS: Performed stx 8 -> stx 10 upgrade activation. Observed that the
      task that install the ssl cert are executed, and that the file
      creation date timestamp matches the tasks execution timestamp.

Story: 2009811
Task: 51152

Change-Id: I726236b2859fc228769f70230eb4d5ea377b6241
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>