Fix rook-ceph restore (BnR)
During the restore playbook, with the "wipe_ceph_osds=true" flag,
the "Delete all resources from the rook-ceph namespace failed"
task may fail. This happens between getting the list of CRDs and
actually deleting any resources from that CRDs in the namespace
rook-ceph, the CRD may no longer exist. This will cause the return
code to be non-zero and the playbook will fail. To prevent this
fom happening, "ignore_erros" was added.

Additionally, some unexpected behavior was found when the target
host is the last to be unlocked in an "ONLY_OSD" scenario:

- After recovery, the monitor that was moved is not “working”:
  This happens because the monitor data directory was not being
  removed when moving it back.

- The OSDs on the other hosts do not go up:
  This happens because after recovery, the OSD keyrings are
  overridden. Therefore, to resolve this issue, keyrings are
  updated only after the recovery process is complete.

- Some monitor may have its name incremented:
  This happens because as the recovery process has not yet
  started, the operator is monitoring/managing the monitors,
  which may result in attempts to create a new monitor.
  To prevent this from happening, the "ceph.rook.io/do-not-
  reconcile" label was set during playbook execution, rather
  than only running in the recovery job.

Test Plan:
- PASS: Configure Standard with OSDs on controller-1
        and worker-0, and monitors on the other hosts.
- PASS: Perform B&R
- PASS: After restore, unlock controller-1 only after
	all hosts have been unlocked.
- PASS: Check ceph status after recovery

Closes-Bug: 2084204

Change-Id: I6ce5bad3697e7cfc49b2d35d05339964d549f4fe
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>