Fix bootstrap replay issue on network reconfiguration
In [1], some of the load endpoints were removed from sysinv, breaking
the bootstrap replay functionality after an incomplete bootstrap,
because it tries to list the loads using sysinv's load list method,
which was was deprecated.
In addition to fixing this issue, this commit also updates the network
reconfiguration logic in populate_initial_config.py to enable changing
the entries in localhost.yml after an incomplete bootstrap. Previously,
if a network overlapped with another, it wasn't possible to update their
entries to fix the issue as the address pool was not removed on a
subsequent bootstrap, causing all retries to fail.

Test plan:
1. PASS: Deploy an AIO-SX.
2. PASS: Deploy a DC with two subclouds.
3. PASS: Bootstrap an AIO-SX with overlapping networks for the cluster
   service subnet, update the network configuration and successfully
   bootstrap the system.
4. PASS: Bootstrap a subcloud with overlapping networks for the external
   oam secondary subnet, update the network configuration and
   successfully bootstrap the system.
5. PASS: Execute bootstrap replay after an incomplete bootstrap that has
   the host created and confirm that the routes are correctly deleted.
6. PASS: Bootstrap a dual stack system with the secondary address
   overlapping with another network and verify that, on a replay, the
   entry is correctly updated.
7. PASS: Bootstrap a dual stack system with the secondary address
   overlapping with another network, replay the bootstrap removing the
   dual stack configuration and verify that all unnecessary entries are
   removed from the database.
8. PASS: Run system show and verify that the load info is correctly
   displayed.
9. PASS: Run system host-show and verify that the load info is correctly
   displayed.

[1] https://review.opendev.org/c/starlingx/config/+/921564

Closes-bug: 2084484

Change-Id: Icb6ec7a874553ea6873a53037fbf3666baed111f
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>