K8S version check before backup
When the customer upgrades the equipment to a new StarlingX version and
initiates a system backup, the code verifies whether the current
k8s version is the latest supported one by the new StarlingX release.
If the k8s version is outdated, the user will be prompted to upgrade
it.
This verification ensures that backups are compatible with the new
environment, preventing errors during the restore process due to an
outdated k8s version.

The user can bypass this verification by setting the override flag:
ignore_kube_version_check_for_backup=true

Note: With the following changes, StarlingX upgrades will only proceed
if the k8s version is the latest supported one by the previous release:
https://review.opendev.org/c/starlingx/config/+/873443
https://review.opendev.org/c/starlingx/distcloud/+/873445

Tests done:
- Upgrade subcloud and execute the backup
- Upgrade AIO-SX and execute the backup
- Change the kube_supported_versions.py to return
	different versions and check if backup will fail
- Change the kube_supported_versions.py to return
	empty list and check if backup will fail
- Change the kube_supported_versions.py to return
	different versions and use the override option
	to continue the backup
        ( flag: ignore_kube_version_check_for_backup )
- AIO-DX fresh install, and do backup and Restore

Closes-Bug: 2084701
Change-Id: Ib2ece4141a3732674ec87614da10b9249f6dbbab