Configure systemd CPUShares for Kubernetes bootstrap
This configures the bootstrap version of systemd service DropIn
overrides for containerd.service and kubelet.service to use
k8splatform.slice.

bringup-essential-services/tasks/bringup_helm.yml was modified to run
'helm repo index' as root instead of www user. This prevents playbook
failing due to permissions problem when it tries to access
.config/helm/registry/config.json, even though the file doesnt exist.
Same permission issue seen after B&R. This was hit repeatedly with the
new CPUShares.

Backup and restore was updated to include systemd DropIn config files,
and contents of /root so we retain toprc. This was required to
retain the k8splatform.slice and customizations, and other overrides
that were missing from the backup.

This is part of an overall set of adjustments are required for systemd
cgroups CPUShares, CPUQuota, and AllowedCPUs for key system services.
This will improve latency of Kubernetes critical components, and
throttles lesser important services.

Partial-Bug: 2084714

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/932569

TEST PLAN:
AIO-SX, AIO-DX, Standard, Storage, DC:
- PASS: Fresh install
- PASS: bootstrap: Verify K8S services run under k8splaform.slice,
        i.e., systemctl status k8splatform.slice

AIO-SX:
- PASS: ansible-playbook replay
- PASS: B&R: verify toprc for root and sysadmin
- PASS: application removal/deletion to test helm lifecycle
  e.g.,
  system application-upload /usr/local/share/applications/helm/metrics-server-24.09-57.tgz
  system application apply metrics-server
  system application remove metrics-server
  system application delete metrics-server

AIO-SX, AIO-DX:
- PASS: B&R: verify systemd DropIn config files
- PASS: B&R: verify toprc for root and sysadmin
- PASS: K8S orchestrated Upgrade 1.24 - 1.29
- PASS: Platform USM Upgrade and pre-activation rollback;
        verify systemd k8splatform.slice + DropIn config

AIO-DX:
- PASS: B&R: Restore with backup taken on controller-1

Change-Id: Ibd951dc09c93254ab6669f9a22093757b1295127
Signed-off-by: Jim Gauld <James.Gauld@windriver.com>