Handle concurrency of multiple prestage instances
Multiple instances of prestage for-sw-deploy are generating
errors when trying to simultaneously access the resource
that is responsible for obtaining the commits from metadata
files.

This occurs because an if/else block was mistakenly removed,
according to the following commit:

https://review.opendev.org/c/starlingx/ansible-playbooks/+/928467

This commit reverts the change to the if/else block to handle
concurrency appropriately.

Test Plan:
PASS: - Bring up a DC#1 with 24.09 and multiple subclouds.
      - Upload and deploy patch 24.09.1 on System Controller.
      - Create a prestage strategy for software deploy.
      - Apply the strategy
      - Verify that the 24.09.1 patch is prestaged on all subclouds.

Story: 2010676
Task: 51037

Change-Id: Iae5f569c87b5a7317efc69c3f6599c0c3b99bfcc
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>