Fix rook-ceph recovery (B&R)
In some scenarios, after recovering rook-ceph, the ceph status
remained as if the cluster was empty:

  cluster:
    id:     1c6688fc-58b8-4055-abc7-c1c730a88608
    health: HEALTH_OK

  services:
    mon: 1 daemons, quorum a (age 4m)
    mgr: no daemons active
    osd: 0 osds: 0 up, 0 in

  data:
    pools:   0 pools, 0 pgs
    objects: 0 objects, 0 B
    usage:   0 B used, 0 B / 0 B avail
    pgs:

After analyzing the code and logs, it was possible to observe
changes caused by race conditions in rook-ceph secrets and
configmaps. To prevent this from happening, the rook-ceph
operator is paused while the recovery process patches secrets
and configmaps.

Furthermore, it was observed that when we do not have OSD and
monitor on the same host, the mgr and mds were on the host
where the recovery was performed (which only has OSD). To fix
this, the respective deployments are restarted, thus placing
them on the correct hosts.

Finally, the logs of all jobs/pods used during recovery are
saved to /var/log/ceph/restore.log on the target host, which
is where the recovery job runs.

Test Plan:
- PASS: Configure Standard with OSDs on controller-1
        and worker-0, and monitors on the other hosts.
- PASS: Perform B&R
- PASS: After restore, unlock controller-1 only after
	    all hosts have been unlocked.
- PASS: Check ceph status after recovery

Closes-Bug: 2086473

Change-Id: If2cd7186eaa4510ce55bcf9b10a9536d76e366c4
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>