Add playbooks to update cephfs k8s PV parameters
Adding new playbooks as helpers to update kubernetes PVs that are
provisioned by cephfs Storage Class and do not have the parameter
'kernelMountOptions: recover_session=clean'. This parameter is
required for the cephfs kernel driver to remount the volume when there
is a connection issue or a client eviction issued by the Ceph
monitoring script.

These playbooks are supposed to be engineering tools to avoid
redeploying applications, forcing recreating the PVCs/PVs.

Test-Plan:
  PASS: Create a deployment that has at least one pod that uses a
    cephfs PVC. Apply the deployment and after the pod is in Running
    state, run the ansible playbook. Check if the parameter is added
    to the storage class cephfs and to the PVs.

Closes-bug: 2085648

Change-Id: I080ee47cc4d7f60e99a29202128560531143abef
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>