Avoid install_uuid mismatch error message during upgrade from N-1 to N release
When a subcloud is deployed with an N-1 release, a file called
install_uuid is generated within the N-1 release feed repo directory.

That file contains the ID expressed in UUID format of the controller
corresponding to the subcloud.

When release N is prestaged for upgrade from the central, release N
will have the install_uuid file with the UUID value corresponding
to the controller of the central, differing from the original value
of the subcloud.

If the UUID does not match, the software agent returns a mismatch
value error.

This commit adds changes to replace the install_uuid file from the
release N-1 feed directory to the release N directory.

Test Plan:
PASS - Bring up a DC with 24.09 and a subcloud with N-1 with USM
       functionality.
     - Prestage the subcloud for software deploy.
     - Verify that the 24.09 release is prestaged on the subcloud.
     - Compare the install_uuid file between the two feed
       directories of release N and N-1. The UUID should be the
       same.
     - Upgrade the subcloud for prestaged N release.
     - Verify that the upgrade is completed and that UUID mismatch
       messages are not displayed in the software log.

Story: 2010676
Task: 51330

Change-Id: I9b571264778e40be013cd8c296d18d2b873af3f0
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>