Update mgmt network during enrollment
This commit adds support to update the primary stack management
network during subcloud enrollment in the case:
1. the system is simplex
2. no admin network configured

As the reconfiguration of the sencondary management network was not
properly tested, it is unknown if potential issues if update two
stacks together. Further, the communication between the system
controllers and the subcloud is based on the primary stack. This
commit doesn't include the task to update the secondary stack
management network.

Test plan:
1. Passed - update management network during enrollment on SX subcloud
w/o admin network configured, verify the management network updated
and the subcloud can be online and in-sync post unlock and managed.
2. Passed - update management floating address only during enrollment
on SX subcloud, verify the floating address is added into the address
pool, verify the subcloud can be online and in-sync post unlock and
managed.
3. Passed - subcloud enrollment against a DX system with management
network update(in bootstrap values), verified the update of the
management network is skipped.
4. Passed - subcloud enrollment against a SX subcloud with both admin
network and management network update, verifed the update of the
management network is skipped, the admin network updated. The subcloud
can be online and in-sync post unlock and managed.
5. Passed - rehomed a subcloud to the DC system with & without admin
network configured.

Depends-on: https://review.opendev.org/c/starlingx/config/+/936201
Closes-bug: 2089596

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: If9977c0c8e69c7de940a436c31e960c3f93d7206