Add Openabo Backup and Restore
Adding backup and restore for openbao. As openbao is a fork of hashicorp
vault, it uses identical procedure to make backup and restore itself.
The process used here is same as the one used for hashicorp vault. The
code is copied over to keep the code separated from the vault version.
This is to keep the work simple once vault backup and restore is
removed.

Test Plan:
PASS    Openbao standalone backup is successful without errors
PASS    Openbao platform backup is successful without errors
PASS    Openbao standalone restore is successful without errors

Story: 2011244
Task: 51419

Change-Id: I49caf7f300563d511a5f609b40defd79420f110c
Signed-off-by: Tae Park <tae.park@windriver.com>