Add retry attempts when retrieving system controller subnet
The creation of the oam network during subcloud rehoming involves
applying the runtime manifest. This process is performed
asynchronously. Since dnsmasq is restarted during this process, it is
necessary to retry accessing the system controller information in case
of failures in the next step of the Ansible playbook.

Test Plan
=========

[PASS] Execute rehoming to a different system controller using the
rehoming playbook. It should not fail. Check if the entries for
system-controller-oam-subnet and system-controller-subnet have been
correctly updated in the addrpool table of the subcloud.

[PASS] Simulate a dnsmasq failure by running the "sm-unmanage service
dnsmasq" command and killing the dnsmasq process before fetching
system controller information from addrpool table. Ansible should stop
the rehoming process and report a failure.

Closes-Bug: 2089932

Change-Id: I7d92c24133e043c2d6ed7118454183f3663358a0
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>