Update Multus Kubeconfig after token expiry
This issue was reported to have occured once the ServiceAccount token
expires after one year (default refresh value). This was reported
on upstream Multus as well.

This change adds a sidecar "watcher" container to the multus deployment
where it will check if the ServiceAccount token was updated by the
kube-api server, and update the Multus Kubeconfig file if not already
updated automatically.

Test:
- Build successful
- DX and SX deploy successful
- Force the ServiceAccount token to expire every hour
  -> Check if the kubeconfig file was updated correctly
  -> Check if the dependent pods are still stuck in "ContainerCreating"
     after the config update

Closes-Bug: 2090964

Change-Id: Ib2b3d73b8dfffb491ba9cae563261b511751b06e
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>