Add tasks to support remote restore
In remote restore, the tasks to retrieve "system-local-ca" are
failing on a standalone system due to the "connection: local"
connection which are added for a 'subcloud' to retrieve the
data from systemcontroller. This fix addresses this issue by
adding a separate block of tasks to support standalone systems.

Test Cases:
PASS: Perform legacy restore from remote workstation on standalone
      system and verify it is successful
PASS: Perform optimized restore and verify its successful
PASS: Deploy systemcontroller and bootstrap subcloud and verify it
      is successful.
PASS: Deploy standalone system and verify it is successful
PASS: Perform legacy restore locally and verify it is successful

Closes-bug: 2091167

Change-Id: Ia1c53741f72b27d8f0c2a31592924f1e8bab0f5f
Signed-off-by: amantri <ayyappa.mantri@windriver.com>