Add cert renewal in enrollment init timeout
In some cases, the runtime configuration that is supposed to
install the ssl certificate containing the new OAM IP in
/etc/ssl/private/server-cert.pem is not working during
enrollment, when the REST API/GUI certificate is updated.

This change adds a remediation for the enrollment failure when
this happens, triggering another renewal of the REST API/GUI
certificate.

It also increases the retries before trying to renew the cert
from 30 to 45, and the timeout before the sysinv API is ready
after the first reboot from 900s to 1020s.

Test plan:
PASS: Verify that renewal tasks are triggered when REST API/GUI
      cert is not valid (curl command fails).
      Verify that fail message is shown when certificate doesn't
      exist.
PASS: Enroll system as subcloud.

Closes-bug: 2091437

Change-Id: Ibd549cc8c2d0f07db4ac6d7889803e806a6bb7a2
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>