Adjust timeout for calico readinessProbe
The timeout value for the readinessProbe of calico kube controller was
increased. This adjustment was necessary to avoid issues during stress
testing, ensuring that the component can properly handle high-load
situations and prevent premature failure in readiness checks.

Test Plan
=========

[PASS] Execute fresh install and verify that the readinessProbe timeout
parameter for the calico-kube-controllers pod has been set to 5 seconds.

Closes-Bug: 2092582

Change-Id: I092ce7e976a5e19b132fc5498c67c6997bc596b2
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>