Show the prestaged versions list for the prestage for-sw-deploy
When the prestage is executed regardless of whether
it is for-install or for-sw-deploy, upon completion
it should display the list of prestaged versions.

Currently, when prestage is for-install, the list
is displayed correctly. However, when the prestage
is for-sw-deploy, the list comes empty, not
providing the correct information.

This commit adds code to support both for-install
and for-sw-deploy cases to display the list of
prestaged versions correctly.

Test Plan:
PASS: - Bring up a DC with 24.09 with at least 1 subcloud
      - Upload and deploy patch 24.09.1 System Controller
      - Prestage the subcloud for-install
      - Verify that the release is prestaged on the subcloud
      - Verify that the prestaged version is shown in
        dcmanager subcloud show output

PASS: - Prestage the subcloud for-sw-deploy
      - Verify that the release is prestaged on the subcloud
      - Verify that the prestaged version is shown in
        dcmanager subcloud show output

Story: 2010676
Task: 51513

Change-Id: I6568a3437d151ce16e9f1b195ddc08b135a6ec00
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>