Enroll certificates verification
This commit compares the system-local-ca issuer certificates
installed in the system controller and in the subcloud during
init enroll operation and fails the playbook if one of the
certificates are not the same between system controller and
subcloud.

Test Plan:

PASS: Factory install a subcloud without certificates and enroll
the subcloud in the system controller. Verify that the enroll init
playbook fails.

PASS: Factory install a subcloud without certificates. Remove any
previous stored keys in root user 'known_hosts'. Verify that the
subcloud key is added and the enroll fails due to a mismatch
between the certificates.

PASS: Factory install a subcloud with the same certificates as the
system controller. Remove any previous stored keys in root user
'known_hosts' and enroll the subcloud. Verify that the subcloud
enrollment completes without errors.

Closes-bug: 2094925

Depends-on: https://review.opendev.org/c/starlingx/distcloud/+/939503

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: Ie6de5fe9272a47a3a8ae41e733d5437642c8368b