Fix default end address values for IPv6
Ensure default start and end values for a network range cover the entire
available range. For IPv6, include the entire range, while for IPv4,
exclude the last address due to its use as a broadcast address.

Test Plan
=========

[PASS] Bootstrap dual-stack with IPv6 as primary address. Use the last
IPv6 address of the network for the external_oam_floating_address
parameter, and the second-to-last address for IPv4. Do not
specify the values for external_oam_end_address and
external_oam_start_address in the localhost.yml file, so that the
default values can be applied. The OAM address pool range must include
the last address of the network for IPv6, but must not include the last
address of the network for IPv4.

[PASS] Repeat the previous test by reversing the configuration,
setting IPv4 as the primary address and IPv6 as the secondary address.

[PASS] Bootstrap dual-stack with IPv6 as primary address. Apply default
values for cluster_pod_end_address, cluster_service_end_address,
admin_end_address, management_multicast_end_address. The address pool
range must include the last address of the network for IPv6, but must
not include the last address of the network for IPv4 for all networks.

[PASS] Run the subcloud enrollment process using a dual-stack
configuration. Use the last network address for IPv6 as the
admin_end_address and the default value for the
external_oam_end_address parameter. For IPv6, the last network address
must be assigned as the end address for both OAM and admin. For IPv4,
the second-to-last address in the network should be used as the end
address for all network types when not otherwise specified.

Closes-Bug: 2095524

Change-Id: If42795b2a88f5b8238fc116b067127b2bdf40376
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>