Fix rook-ceph backup and restore
In some scenarios, the rook-ceph recovery job may get stuck
when recovering cephfs, this happens because there is no mgr
active at that time. Furthermore, when it gets stuck in this
part, there is a timeout that is exceeded, however the
playbook does not fail and the next task is executed as if
the rook-ceph recovery had been done successfully.

To fix this, mgr is scaled to replicas=1. As a result, it was
necessary to make some changes to the code to follow this new
flow. Regarding the timeout, it is now being checked and when
it occurs, the playbook will fail the "Recover rook-ceph" task.

Additionally, a return code check has been added to all ceph
commands to ensure they are executed correctly, and fail if
any command returns anything other than 0.

Test Plan:
- PASS: optimized AIO-SX B&R
- PASS: optimized AIO-SX B&R with wipe_ceph_osds flag
- PASS: legacy DX B&R
- PASS: legacy DX B&R with wipe_ceph_osds flag
- PASS: legacy STD B&R

Closes-Bug: 2097369

Change-Id: Ia0d6b63acfa36a6eb978f98d53d6a6ff06af7071
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>